mktemp, mktempdir: delete temp files on exit by default
Instead of putting lots of atexit handlers in the list, this uses
a single one and keeps a list of temporary paths to cleanup. If the
list gets long (≥ 4096) it will be scanned for already-deleted paths
when adding a new path to the cleanup list. To avoid running too
frequently, it doubles the "too many paths" threshold after this.